Make comparison and hash functions take const args
Comparison and hash functions are generally assumed to not mutate the compared/hashed elements.
This makes the compiler check that we do not directly mutate in these functions.

This also necessitates changing the *_cmp_fn type in the list macros at `src/libponyrt/ds/list.h`,
and allowing a warning for that line, for which an explanatory comment is added:

```
/* The following disables the duplicate-decl-specifier warning for\
  clang and gcc. The duplicate-decl-specifier warning warns on code like\
  `const char const*x;` which duplicates const, because the code\
  `const char*const x` is generally what is meant.\
  Here we do intend the first meaning, when `elem` is `const char`,\
  so we disable the warning for the affected line */\
  _Pragma("GCC diagnostic push") \
  _Pragma("GCC diagnostic ignored \"-Wduplicate-decl-specifier\"") \
  typedef bool (*name##_cmp_fn)(elem const* a, elem const* b); \
  _Pragma("GCC diagnostic pop")\
```